+++ packages/ui/src/elements/index.ts
@@ -1,2 +1,3 @@
+export * from './onClickOutside'
 export * from './input'
 export * from './settings-overlay'