+++ packages/ui/src/index.tsx
@@ -1,5 +1,4 @@
 export * from './styled'
 export * from './icon'
-export * from './plugin-rows'
 export * from './elements'
 export * from './upload-progress'